requirements.txt
binance-connector
python-dotenv
